% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{clean_names}
\alias{clean_names}
\title{Clean names}
\usage{
clean_names(x)
}
\arguments{
\item{x}{the object with names to be cleaned}
}
\value{
A data frame with names in lowercase, spaces as underscores and brackets and forward slashes removed
}
\description{
Clean names
}
\examples{
\dontrun{df <- clean_name(df)}
}
