% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_featurecodes}
\alias{get_featurecodes}
\title{Geonames feature codes}
\format{{
A data.frame with 11 variables and 4084 rows
\describe{
\item{\code{feature_class}}{a character vector containing the Geonames feature class}
\item{\code{feature_code}}{a character vector containing the Geonames feature code}
\item{\code{feature_type}}{a character vector containing the Geonames feature type}
\item{\code{feature_description}}{a character vector containing the Geonames feature description}
}
}}
\usage{
get_featurecodes()
}
\value{
data frame
}
\description{
Geonames feature codes
}
\details{
The main allCountries table that is the source for the undersea table lists features in terms of feature class and feature code. The featurecodes table concatenates these fields with a stop. This function separates them for joining.
}
\examples{
\dontrun{get_featurecodes()}
}
\references{
Geonames feature code file \url{http://download.geonames.org/export/dump/featureCodes_en.txt} and table \url{http://www.geonames.org/export/codes.html}
}
