% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_vents.R
\name{import_vents}
\alias{import_vents}
\title{Import the latest Interridge Vents Field data}
\usage{
import_vents()
}
\value{
data.frame
}
\description{
Import the latest version of the csv file from the InterRidge Vents Database \url{https://vents-data.interridge.org/ventfields_list_all}, presently in Version 3.4. The function imports the latest version into data-raw. Column names are regularised by replacing spaces with underscores,removing brackets and a download_date column is added. The file is saved to data as \file{vents.rda} and loaded into the global environment as \file{vents}.
}
\details{
For license details see \file{interridge_citation.txt} in data-raw.
}
\examples{
\dontrun{vents <- import_vents()}
}
\references{
Please cite the data as: "Beaulieu, S.E., 2015, InterRidge Global Database of Active Submarine Hydrothermal Vent Fields: prepared for InterRidge, Version 3.3. World Wide Web electronic publication. Version 3.X accessed 2017-09-04, http://vents-data.interridge.org"
}
