% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{pass lhs to rhs and do something}
}
\description{
Like dplyr underwater uses the pipe function, \code{\%>\%}.
}
\examples{
\dontrun{vents \%>\% select(name_id)}
}
