% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_undersea.R
\docType{data}
\name{undersea}
\alias{undersea}
\title{Geonames Undersea Place Names}
\format{A data.frame with 20 variables \code{geonameid}, \code{name}, \code{asciiname}, \code{alternatenames}, \code{latitude}, \code{longitude}, \code{feature_class}, \code{feature_code}, \code{iso}, \code{cc2}, \code{admin1_code}, \code{admin2_code}, \code{admin3_code}, \code{admin4_code}, \code{population}, \code{elevation}, \code{dem}, \code{timezone}, \code{modification_date}, \code{download_date}}
\source{
Extracted from \href{http://download.geonames.org/export/dump/allCountries.zip}{Geonames allCountries file}
}
\usage{
undersea
}
\description{
A data frame with 14,513 Geonames places with feature class U "undersea" extracted from \href{http://download.geonames.org/export/dump/allCountries.zip}{Geonames allCountries.txt} on 2018-01-01. For all feature codes see \href{http://www.geonames.org/export/codes.html}{GeoNames Feature Codes}.
}
\details{
Column names converted to lower case, punctuation and spaces replaced with underscores.
}
\references{
Suggested Citation Geonames Feature Class U \href{http://download.geonames.org/export/dump/allCountries.zip}{Geonames allCountries file}
}
\keyword{datasets}
