% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gebco.R
\name{import_gebco}
\alias{import_gebco}
\title{import gebco from raw}
\usage{
import_gebco(path = NULL)
}
\arguments{
\item{path}{path to a local file (optional)}
}
\value{
data.frame
}
\description{
Imports the raw gebco csv file in data-raw, regularize column names (removes brackets and forward slash, replace space with underscore).
}
\details{
The raw csv file is manually downloaded to data-raw from \url{https://www.ngdc.noaa.gov/gazetteer/} and is Gazeteer version 2.0+13. Download date 2017-12-29.
}
\examples{
\dontrun{
gebco <- import_gebco()
path <- import_gebco(path = "pathtodfile")
}
}
\references{
Please cite as "IHO-IOC GEBCO Gazetteer of Undersea Feature Names, www.gebco.net"
}
