% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe}
\arguments{
\item{lhs, rhs}{pass lhs to rhs and do}
}
\description{
Like dplyr underwater uses the pipe function, \code{\%>\%}. Copied from ggvis pipe.R
}
\examples{
\dontrun{mtcars \%>\% select(mpg)}
}
